Web-Scraping/WebsiteScrape/requirements.txt
Scrapy>=2.11
pymongo>=4.6
python-dotenv>=1.0